src/tool/mcp_tools/computer_use_tool/pyproject.toml
[tool.poetry]
name = "computer_use_tool"
version = "0.1.0"
description = "MCP container exposing computer use automation via FastMCP"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"

# MCP server
fastmcp = "^2.0.0"
loguru = "^0.7.2"

# Computer use automation
ipython = "^8.29.0"
python-dotenv = "^1.0.1"
gradio-client = "^1.5.0"
huggingface-hub = "^0.24.0"
pillow = "^11.0.0"
openai = "^1.59.5"
anthropic = "^0.44.0"
numpy = "^2.3.5"

[tool.poetry.scripts]
start = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
